requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.12.1
python-jose[cryptography]==3.3.0
bcrypt==3.1.0
passlib==1.7.4
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
coverage==7.3.2
openai-whisper==20250625
torch==2.7.1
torchaudio==2.7.1
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
pydantic_settings=2.10.1
python-multipart=0.0.20
redis=6.2.0